refactor(api): move distributed cache file and test
